requirements.txt
streamlit==0.88.0
transformers==4.11.3
torch==1.9.0
tensorflow==2.6.0
Pillow==8.3.2
click<8.0,>=7.0
numpy~=1.19.2
six~=1.15.0
